JavaDoc and code cleanup

+++ src/main/java/com/rdebokx/formica/core/Ant.java
@@ -4,18 +4,35 @@ import com.rdebokx.formica.execution.Configuration;
 
 import java.util.List;
 
+/**
+ * Ant class, representing an ant that can pick up and drop DataPoints when moving from bucket to bucket, based on similarity of (or to) the DataPoints in the bucket.
+ */
 public class Ant {
 
+  /**
+   * The Colony that this ant is currently part of.
+   * Ant will use the randomizer and distanceCalculator of this colony.
+   */
+  private final Colony colony;
 
-  private Colony colony;
-
+  /**
+   * The DataPoint that the ant is currently carrying.
+   */
   private DataPoint payload;
 
-
+  /**
+   * Constructor, constructing an Ant for the provided Colony.
+   * @param colony The Colony that this Ant belongs to.
+   */
   public Ant(Colony colony){
     this.colony = colony;
   }
 
+  /**
+   * Move this Ant to the given bucket. Upon moving this bucket, the probabilistic function drop or pickUp will be called, based on whether this ant is carrying a payload or not respectively.
+   * Note that after this function finished, the ant will have no notion of which bucket it is currently at.
+   * @param nextBucket The bucket that this ant is moving to.
+   */
   public void move(List<DataPoint> nextBucket) {
     if(payload == null){
       pickUp(nextBucket);
@@ -25,13 +42,19 @@ public class Ant {
   }
 
   /**
-   * TODO javadoc. Probabilistic function for picking up.
-   * @param nextBucket
+   * Probabilistic function for picking up a new DataPoint from the given bucket. This function will randomly select a DataPoint from the given bucket.
+   * With a chance P, which is based on the similarity of this DataPoint to the rest of the bucket, this DataPoint will then be removed from the bucket and carried by this ant as payload.
+   * The less similar the selected DataPoint is to the other DataPoints in the bucket, the higher the chance that this DataPoint will be picked up by this ant.
+   * Will use the randomizer of the Colony that this ant belongs to, to evaluate whether or not to pickup the selected DataPoint.
+   * @param nextBucket The bucket that this ant has moved to and a DataPoint may or may not be picked up from.
    */
-  public void pickUp(List<DataPoint> nextBucket){
+  protected void pickUp(List<DataPoint> nextBucket){
+    if(payload != null){
+      throw new RuntimeException("Ant is already carrying payload.");
+    }
     if(!nextBucket.isEmpty()) {
       int candidateIndex = colony.randomizer.nextInt(nextBucket.size());
-      double distanceToBucket = distanceToBucket(nextBucket.get(candidateIndex), nextBucket, Configuration.BASIC_PICKUP_PROB);
+      double distanceToBucket = nextBucket.get(candidateIndex).avgDistanceToBucket(nextBucket, colony.distanceCalculator, Configuration.BASIC_PICKUP_PROB);
 
       if (performAction(distanceToBucket)) {
         payload = nextBucket.remove(candidateIndex);
@@ -39,32 +62,30 @@ public class Ant {
     }
   }
 
-  public void drop(List<DataPoint> nextBucket){
-    if(performAction(1 - distanceToBucket(payload, nextBucket, 1 - Configuration.BASIC_DROP_PROB))){
+  /**
+   * Probabilistic function for dropping the current payload that this ant is carrying in the given bucket.
+   * With a chance P, which is based on the similarity of the current payload to the DataPoints in the given bucket, the current payload will be dropped by the ant and added to the given bucket.
+   * The more similar the current payload is to the other DataPoints in the bucket, de higher the chance that the current payload will be added to the bucket.
+   * Will use the randomizer of the Colony that this ant belongs to, to evaluate whether or not to drop the current payload.
+   * @param nextBucket The bucket that his ant may or may not its current payload in.
+   */
+  protected void drop(List<DataPoint> nextBucket){
+    if(payload == null){
+      throw new RuntimeException("Ant has no payload to drop.");
+    }
+    if(performAction(1 - payload.avgDistanceToBucket(nextBucket, colony.distanceCalculator,1 - Configuration.BASIC_DROP_PROB))){
       nextBucket.add(payload);
       payload = null;
     }
   }
 
-  public double distanceToBucket(DataPoint candidate, List<DataPoint> bucket, double defaultDistance){
-    double totalDistance = 0;
-    int distancesCount = 0;
-    for(DataPoint dataPoint : bucket){
-      if(dataPoint != candidate){
-        totalDistance += colony.distanceCalculator.distance(candidate, dataPoint);
-        distancesCount++;
-      }
-    }
-    return distancesCount > 0 ? totalDistance / distancesCount : defaultDistance;
-  }
-
   /**
-   * TODO javadoc.
-   * Probabilistic function.
-   * @param probability double value between 0 and 1. The higher the number, the higher the chance of true being returned.
-   * @return
+   * Probabilistic helper function to determine whether or not to perform an action, based on the probability P.
+   * @param p Probability P that true will be returned.
+   *          Double value between 0 and 1, where 0 gives a 0% chance that this function will return true and 1 will give a 100% chance that this function will return true.
+   * @return true with probability P.
    */
-  private boolean performAction(double probability) {
-    return colony.randomizer.nextDouble() < probability;
+  private boolean performAction(double p) {
+    return colony.randomizer.nextDouble() < p;
   }
 }
